% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiCalls.R
\name{getResults}
\alias{getResults}
\title{Get simulation job results.}
\usage{
getResults(job_id, type = "output")
}
\arguments{
\item{job_id}{Simulation's id whose results the user wants to retrieve (as Integer)}

\item{type}{Type of results the user is interested in (String value, one of: "output", "pathway_output", "nodes_output"; Default: "output")}
}
\value{
The function returns the requested results as data frame.
}
\description{
Let the user retrieve all results from a specified simulation job.
}
\details{
The user can choose from 3 different results, specified by the function parameter \emph{type}.\cr
\itemize{
\item{\emph{output}}{ : results will be given in "raw" format, an "all-in-one" solution that shows every evaluation from PHENSIM (Activity-Score, P-value, ...) for each node (biological element) from every pathway;}
\item{\emph{pathway_output}}{ : matrix-form auxilary details showing all random values generated during the simulation process, given by pathway;}
\item{\emph{nodes_output}}{ : matrix-form auxilary details showing all random values generated during the simulation process, given by nodes (biological elements);}
}
}
\examples{
df <- getResults(3777, type="pathway_output")
}
