% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiCalls.R
\name{infoJob}
\alias{infoJob}
\title{Prints more details for queried job}
\usage{
infoJob(job_id)
}
\arguments{
\item{job_id}{Simulation's id (as Integer)}
}
\value{
Details for queried job
}
\description{
Gives all details about the requested job id. Typical work-flow would be calling listJobs (check "?listJobs") and calling infoJob passing the job_id.\cr
The method gives information about creation data, job's status, id, name, organism, and simulation parameters.\cr
For more details check PHENSIM docs at \url{https://phensim.tech/docs/api}
}
\examples{
infoJob(691)
}
