% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiCalls.R
\name{jobParameters}
\alias{jobParameters}
\title{Get job simulation parameters}
\usage{
jobParameters(job_id, save = 0, view = 1)
}
\arguments{
\item{job_id}{Simulation's id whose parameters will be retrieved (as Integer);}

\item{save}{Set this flag to 1 to save job parameters as local file (default 0);}

\item{view}{This flag is set to 1 by default and will always print data (can be set to 0 otherwise);}
}
\value{
This function prints the requested info and also returns the data as String (is var assignable).
}
\description{
Retrieve parameters for specific job simulation.
}
\details{
Requests and prints the parameters file for a specific job, as inputted during submission. This file should contain only the nodes' expression indication.\cr
This method requires the simulation's id, accessible from listJobs() details.\cr
Note: this method also gives the possibility to save locally the file.\cr

For more details check PHENSIM docs at \url{https://phensim.tech/docs/api}
}
\examples{
jobParameters(691)

}
