% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiCalls.R
\name{listJobs}
\alias{listJobs}
\title{Lists all jobs available to the current user}
\usage{
listJobs()
}
\value{
Printed info
}
\description{
Lists all jobs available to the user at that moment, printing the jobs' IDs, names and current states \cr
For more details check PHENSIM docs at \url{https://phensim.tech/docs/api}
}
