% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiCalls.R
\name{serviceInfo}
\alias{serviceInfo}
\title{Test APIs service status}
\usage{
serviceInfo()
}
\value{
If the token is correctly set up, prints the user account info
}
\description{
Checks if APIs token is correctly set up. Furthermore, prints all info about the user account registered for that token.\cr
For more details check PHENSIM docs at \url{https://phensim.tech/docs/api}
}
