% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apiCalls.R
\name{submitJob}
\alias{submitJob}
\title{Interactive procedure to submit a new simulation}
\usage{
submitJob()
}
\value{
Submitting a new simulation job returns the APIs call response with all details.
}
\description{
Starts an interactive procedure to submit a new simulation job in PHENSIM\cr
}
\details{
The procedure guides the user through the steps required to fill all simulation fields.\cr
In each step the user is asked to insert a value or select a file to set that specific simulation parameter\cr
There are 17 parameters the user can set during the procedure:\cr
\itemize{
\item{\emph{name}}{ : the name that wil be given to the simulation;}
\item{\emph{organism}}{ : the organism on which perform the simulation (using KEGG accession number);}
\item{\emph{epsilon}}{ : a numeric value to determine non-expressed node;}
\item{\emph{seed}}{ : seed value for RNG to allow reproducibility;}
\item{\emph{fdr}}{ : string value for FDR alrithm, one of: BH, QV, LOC (default: BH);}
\item{\emph{reactome}}{ : boolean value to use reactome with KEGG;}
\item{\emph{fast}}{ : boolean value to use the fast method for the perturbation computation (default: true);}
\item{\emph{miRNAs}}{ : boolean value to enable MITHrIL miRNA enrichment feature;}
\item{\emph{miRNAEvidence}}{ : string value to select miRNA-target interactions (One of: STRONG, WEAK, PREDICTION; Default: STRONG);}
\item{\emph{simulationParametersFile}}{ : file of simulation parameters, formatted as a list of entity-regulationtype;}
\item{\emph{enrichmentDatabaseFile}}{ : file of enrichment database;}
\item{\emph{filter}}{ : filter for the enrichment database;}
\item{\emph{nonExpressedNodesFile}}{ : file of non-expressed nodes file;}
\item{\emph{knockoutNodesFile}}{ : file of knocked-out nodes;}
\item{\emph{customNodeTypesFile}}{ : file of custom node types;}
\item{\emph{customEdgeTypesFile}}{ : file of custom edge types;}
\item{\emph{customEdgeSubtypesFile}}{ : file of custom edge subtypes;}
}

NOTE: all parameters are OPTIONAL (default value will be used eventually) except for \emph{name}, \emph{organism} and \emph{simulationParametersFile}, they have to be set otherwise the procedure will fail.\cr
Furthermore, the \emph{submit} parameter is automatically set to 1, so the simulation will be automatically submitted once the procedure in done.\cr


For more details check PHENSIM docs at \url{https://phensim.tech/docs/api}
}
